+++ lib/src/annotations/squadron_service.dart
@@ -1,4 +1,7 @@
 import '../pool/worker_pool.dart';
+import 'target_platform.dart';
+
+const localService = SquadronService.local();
 
 /// Annotation for service classes to be wrapped as workers.
 class SquadronService {
@@ -6,7 +9,8 @@ class SquadronService {
     this.pool = true,
     this.targetPlatform = TargetPlatform.all,
     String? baseUrl,
-  }) : baseUrl = baseUrl ?? '';
+  })  : baseUrl = baseUrl ?? '',
+        local = false;
 
   const SquadronService.web({bool pool = true, String? baseUrl})
       : this(pool: pool, targetPlatform: TargetPlatform.web, baseUrl: baseUrl);
@@ -14,9 +18,19 @@ class SquadronService {
   const SquadronService.vm({bool pool = true})
       : this(pool: pool, targetPlatform: TargetPlatform.vm);
 
+  const SquadronService.local()
+      : pool = false,
+        local = true,
+        targetPlatform = TargetPlatform.all,
+        baseUrl = '';
+
   /// Controls code generation of a [WorkerPool] exposing the target service class.
   /// `true` by default.
   final bool pool;
+
+  /// Controls code generation of a [LocalWorkerClient] exposing the target service class.
+  /// `false` by default.
+  final bool local;
 
   /// Controls code generation of a entry points for various platforms.
   final int targetPlatform;
@@ -24,19 +38,4 @@ class SquadronService {
   /// For Web-based workers, indicates the [baseUrl] where the Web Worker will
   /// be exposed in production.
   final String baseUrl;
-}
-
-final class TargetPlatform {
-  static const vm = 1;
-  static const js = 2;
-  static const wasm = 4;
-
-  static const web = js | wasm;
-  static const all = vm | js | wasm;
-}
-
-extension TargetPlatformExt on int {
-  bool get hasVm => (this & TargetPlatform.vm) != 0;
-  bool get hasJs => (this & TargetPlatform.js) != 0;
-  bool get hasWasm => (this & TargetPlatform.wasm) != 0;
 }